fleet_pro/requirements.txt
customtkinter
tkcalendar
matplotlib
pillow
fpdf2
CTkMessagebox
